+++ routes/web.php
@@ -1,8 +1,8 @@
 <?php
 
+use App\Http\Controllers\YoutubeChannelController;
 use Illuminate\Support\Facades\Route;
 
 //Main Page
-Route::get('/', function () {
-    return view('Index');
-})->name('index');
+Route::get('/', [YoutubeChannelController::class, 'index'])
+    ->name('index');