+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Development
 
+* 2026-02-02 -- 0.0.3 Models and Channel Controller
+  * Added YoutubeChannel and YoutubeVideo models
+  * Added YoutubeChannelController for index
+
 * 2026-31-01 -- 0.0.2 Index page and migrations
   * Added Index template
   * Added youtube channel migration